Require doctrine/mongodb-odm v2.0

+++ tests/Document/TestCase.php
@@ -9,21 +9,27 @@
 
 namespace olvlvl\DoctrineGenerators\Document;
 
-use Doctrine\MongoDB\Connection;
 use Doctrine\ODM\MongoDB\Configuration;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\Mapping\ClassMetadataFactory;
-use Doctrine\ODM\MongoDB\Mapping\Driver\SimplifiedYamlDriver;
+use MongoDB\Client;
+use olvlvl\DoctrineYamlDriver\MongoDB\SimplifiedYamlDriver;
 use function getenv;
 use function realpath;
 
 class TestCase extends \PHPUnit\Framework\TestCase
 {
+    use EnsureDirectory;
+
     private $cacheDir;
 
     public function setUp()
     {
         $dir = realpath(__DIR__ . '/../sandbox');
+
+        if (!$dir) {
+            throw new \RuntimeException("sandbox directory is missing.");
+        }
 
         if (file_exists($dir)) {
             shell_exec("rm -rf $dir/*");
@@ -36,10 +42,13 @@ class TestCase extends \PHPUnit\Framework\TestCase
 
     protected function createDocumentManager(Configuration $config): DocumentManager
     {
-        return DocumentManager::create(
-            new Connection(getenv('MONGODB_SERVER'), [], $config),
-            $config
+        $client = new Client(
+            'mongodb://'.getenv('MONGODB_SERVER'),
+            [],
+            [ 'typeMap' => DocumentManager::CLIENT_TYPEMAP ]
         );
+
+        return DocumentManager::create($client, $config);
     }
 
     protected function createConfig(): Configuration
@@ -48,15 +57,15 @@ class TestCase extends \PHPUnit\Framework\TestCase
         $cacheDir = $this->cacheDir;
 
         $config = new Configuration();
-        $config->setProxyDir("$cacheDir/Proxies");
+        $config->setProxyDir($this->ensureDirectory("$cacheDir/Proxies"));
         $config->setProxyNamespace('Proxies');
-        $config->setAutoGenerateProxyClasses(Configuration::AUTOGENERATE_NEVER);
-        $config->setHydratorDir("$cacheDir/Hydrators");
+        $config->setAutoGenerateProxyClasses(Configuration::AUTOGENERATE_FILE_NOT_EXISTS);
+        $config->setHydratorDir($this->ensureDirectory("$cacheDir/Hydrators"));
         $config->setHydratorNamespace('Hydrators');
-        $config->setAutoGenerateHydratorClasses(Configuration::AUTOGENERATE_NEVER);
+        $config->setAutoGenerateHydratorClasses(Configuration::AUTOGENERATE_FILE_NOT_EXISTS);
         $config->setDefaultDB($database);
         $config->setMetadataDriverImpl(
-            new SimplifiedYamlDriver([ __DIR__ . '/dcm' => __NAMESPACE__ ], '.dcm.yml')
+            new SimplifiedYamlDriver([ __DIR__ . '/dcm' => __NAMESPACE__ ])
         );
 
         return $config;
